Move acceptance tests alongside resources and add ValidateConfig coverage
- Move entity resource acceptance tests to entity/acc_test.go with testdata
- Move entities data source acceptance test to entities/acc_test.go with testdata
- Add PlanOnly acceptance tests for resource ValidateConfig:
  - entity_id mismatch with entity.id field
  - entity_id mismatch with entity_json id field
- Add PlanOnly acceptance tests for data source ValidateConfig:
  - Mixed pagination modes error
  - entity_id + filter conflict via schema validator
- Remove duplicate skipIfUnsupportedEntity helper and entity tests from parent acc_test.go
