+++ internal/kibana/security_entity_store/acc_test.go
@@ -216,120 +216,6 @@ func TestAccDataSourceKibanaSecurityEntityStoreStatus_basic(t *testing.T) {
 	})
 }
 
-func TestAccResourceKibanaSecurityEntityStoreEntity_generic(t *testing.T) {
-	skipIfUnsupportedEntity(t)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { acctest.PreCheck(t) },
-		Steps: []resource.TestStep{
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("create_host"),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttrSet("elasticstack_kibana_security_entity_store_entity.test", "id"),
-					resource.TestCheckResourceAttr("elasticstack_kibana_security_entity_store_entity.test", "entity_type", "generic"),
-					resource.TestCheckResourceAttr("elasticstack_kibana_security_entity_store_entity.test", "entity_id", "generic:acc-test-entity"),
-					resource.TestCheckResourceAttrSet("elasticstack_kibana_security_entity_store_entity.test", "document_json"),
-				),
-			},
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("create_host"),
-				PlanOnly:                 true,
-			},
-		},
-	})
-}
-
-func TestAccResourceKibanaSecurityEntityStoreEntity_updateHost(t *testing.T) {
-	skipIfUnsupportedEntity(t)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { acctest.PreCheck(t) },
-		Steps: []resource.TestStep{
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("create_host"),
-			},
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("update_host"),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttrSet("elasticstack_kibana_security_entity_store_entity.test", "id"),
-					resource.TestCheckResourceAttr("elasticstack_kibana_security_entity_store_entity.test", "entity_type", "generic"),
-					resource.TestCheckResourceAttr("elasticstack_kibana_security_entity_store_entity.test", "entity_id", "generic:acc-test-entity"),
-					resource.TestCheckResourceAttr("elasticstack_kibana_security_entity_store_entity.test", "entity.name", "acc-test-entity-updated"),
-				),
-			},
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("update_host"),
-				PlanOnly:                 true,
-			},
-		},
-	})
-}
-
-func TestAccResourceKibanaSecurityEntityStoreEntity_import(t *testing.T) {
-	skipIfUnsupportedEntity(t)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { acctest.PreCheck(t) },
-		Steps: []resource.TestStep{
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("create_host"),
-			},
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("create_host"),
-				ResourceName:             "elasticstack_kibana_security_entity_store_entity.test",
-				ImportState:              true,
-				ImportStateVerify:        true,
-				ImportStateVerifyIgnore:  []string{"force"},
-			},
-		},
-	})
-}
-
-func TestAccResourceKibanaSecurityEntityStoreEntity_entityJsonConflict(t *testing.T) {
-	skipIfUnsupportedEntity(t)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { acctest.PreCheck(t) },
-		Steps: []resource.TestStep{
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("entity_json_conflict"),
-				ExpectError:              regexp.MustCompile("(?i)conflict|ConflictsWith|Invalid Attribute Combination"),
-			},
-		},
-	})
-}
-
-func TestAccDataSourceKibanaSecurityEntityStoreEntities_basic(t *testing.T) {
-	skipIfUnsupportedEntity(t)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() { acctest.PreCheck(t) },
-		Steps: []resource.TestStep{
-			{
-				ProtoV6ProviderFactories: acctest.Providers,
-				ConfigDirectory:          acctest.NamedTestCaseDirectory("list"),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttrSet("data.elasticstack_kibana_security_entity_store_entities.test", "id"),
-					resource.TestCheckResourceAttrSet("data.elasticstack_kibana_security_entity_store_entities.test", "results_json"),
-					resource.TestCheckResourceAttrSet("data.elasticstack_kibana_security_entity_store_entities.test", "items.#"),
-				),
-			},
-		},
-	})
-}
-
-func skipIfUnsupportedEntity(t *testing.T) {
-	versionutils.SkipIfUnsupported(t, securityentitystore.MinVersion, versionutils.FlavorAny)
-}
-
 func skipIfUnsupported(t *testing.T) {
 	versionutils.SkipIfUnsupported(t, securityentitystore.MinVersion, versionutils.FlavorAny)
 }
