+++ internal/kibana/security_entity_store/entities/testdata/TestAccDataSourceKibanaSecurityEntityStoreEntities_entityIdFilterConflict/entity_id_filter_conflict/main.tf
@@ -1,0 +1,4 @@
+data "elasticstack_kibana_security_entity_store_entities" "test" {
+  entity_id = "generic:acc-test-entity"
+  filter    = "entity.type:generic"
+}
