+++ internal/kibana/security_entity_store/entities/testdata/TestAccDataSourceKibanaSecurityEntityStoreEntities_mixedPaginationError/mixed_pagination/main.tf
@@ -1,0 +1,5 @@
+data "elasticstack_kibana_security_entity_store_entities" "test" {
+  sort_field = "entity.id"
+  sort_order = "asc"
+  filter     = "entity.type:generic"
+}
