+++ internal/kibana/security_entity_store/entity/testdata/TestAccResourceKibanaSecurityEntityStoreEntity_entityJsonIdMismatch/entity_json_id_mismatch/main.tf
@@ -1,0 +1,17 @@
+resource "elasticstack_kibana_security_entity_store" "store" {
+  entity_types = ["generic"]
+}
+
+resource "elasticstack_kibana_security_entity_store_entity" "test" {
+  depends_on = [elasticstack_kibana_security_entity_store.store]
+
+  entity_type = "generic"
+  entity_id   = "generic:acc-test-entity"
+
+  entity_json = jsonencode({
+    id     = "generic:wrong-id"
+    name   = "acc-test-entity"
+    type   = "generic"
+    source = ["terraform-acc-test"]
+  })
+}
